clean up (part 6): adjust variable names to new terminology, etc.
