+++ plugins/org.eclipse.n4js.ide/src/org/eclipse/n4js/ide/imports/ReferenceResolution.java
@@ -10,18 +10,12 @@
  */
 package org.eclipse.n4js.ide.imports;
 
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.List;
-
 import org.eclipse.xtext.resource.IEObjectDescription;
-import org.eclipse.xtext.util.ReplaceRegion;
 
 /**
  * Each instance of this class represents one possible resolution of a {@link ReferenceDescriptor reference given in the
  * N4JS source code}. There may be many such possible resolutions for a single reference in the code. Class
- * {@link ReferenceResolutionHelper} can be used to obtain all possible resolutions for a given reference.
+ * {@link ReferenceResolutionFinder} can be used to obtain all possible resolutions for a given reference.
  */
 public class ReferenceResolution {
 
@@ -40,17 +34,13 @@ public class ReferenceResolution {
 	/** An import that has to be added in case the user chooses this resolution. May be <code>null</code>. */
 	public final ImportDescriptor importToBeAdded;
 
-	/** Text replacements that have to be applied in case the user chooses this resolution. May be empty. */
-	public final List<ReplaceRegion> textReplacements;
-
 	/** Creates an instance. See {@link ReferenceResolution}. */
 	public ReferenceResolution(IEObjectDescription referencedElement, String proposal, String label, String description,
-			ImportDescriptor importToBeAdded, Collection<ReplaceRegion> textReplacements) {
+			ImportDescriptor importToBeAdded) {
 		this.referencedElement = referencedElement;
 		this.proposal = proposal;
 		this.label = label;
 		this.description = description;
 		this.importToBeAdded = importToBeAdded;
-		this.textReplacements = Collections.unmodifiableList(new ArrayList<>(textReplacements));
 	}
 }
